Refactor Rovo Dev webview provider
