+++ src/rovo-dev/performanceLogger.test.ts
@@ -21,7 +21,7 @@ describe('PerformanceLogger', () => {
 
     beforeEach(() => {
         jest.clearAllMocks();
-        performanceLogger = new PerformanceLogger();
+        performanceLogger = new PerformanceLogger('test-instance-id');
 
         // Setup mock analytics client
         mockAnalyticsClient = {
@@ -78,7 +78,6 @@ describe('PerformanceLogger', () => {
 
     describe('promptFirstByteReceived', () => {
         beforeEach(() => {
-            performanceLogger.appInitialized('test-instance-123');
             performanceLogger.sessionStarted('test-session-123');
         });
 
@@ -97,7 +96,7 @@ describe('PerformanceLogger', () => {
                 'api.rovodev.chat.response.timeToFirstByte',
                 measureValue,
                 {
-                    appInstanceId: 'test-instance-123',
+                    appInstanceId: 'test-instance-id',
                     rovoDevSessionId: 'test-session-123',
                     rovoDevPromptId,
                 },
@@ -118,7 +117,6 @@ describe('PerformanceLogger', () => {
 
     describe('promptFirstMessageReceived', () => {
         beforeEach(() => {
-            performanceLogger.appInitialized('test-instance-123');
             performanceLogger.sessionStarted('test-session-123');
         });
 
@@ -137,7 +135,7 @@ describe('PerformanceLogger', () => {
                 'api.rovodev.chat.response.timeToFirstMessage',
                 measureValue,
                 {
-                    appInstanceId: 'test-instance-123',
+                    appInstanceId: 'test-instance-id',
                     rovoDevSessionId: 'test-session-123',
                     rovoDevPromptId,
                 },
@@ -151,7 +149,6 @@ describe('PerformanceLogger', () => {
 
     describe('promptTechnicalPlanReceived', () => {
         beforeEach(() => {
-            performanceLogger.appInitialized('test-instance-123');
             performanceLogger.sessionStarted('test-session-123');
         });
 
@@ -170,7 +167,7 @@ describe('PerformanceLogger', () => {
                 'api.rovodev.chat.response.timeToTechPlan',
                 measureValue,
                 {
-                    appInstanceId: 'test-instance-123',
+                    appInstanceId: 'test-instance-id',
                     rovoDevSessionId: 'test-session-123',
                     rovoDevPromptId,
                 },
@@ -184,7 +181,6 @@ describe('PerformanceLogger', () => {
 
     describe('promptLastMessageReceived', () => {
         beforeEach(() => {
-            performanceLogger.appInitialized('test-instance-123');
             performanceLogger.sessionStarted('test-session-123');
         });
 
@@ -204,7 +200,7 @@ describe('PerformanceLogger', () => {
                 'api.rovodev.chat.response.timeToLastMessage',
                 measureValue,
                 {
-                    appInstanceId: 'test-instance-123',
+                    appInstanceId: 'test-instance-id',
                     rovoDevSessionId: 'test-session-123',
                     rovoDevPromptId,
                 },
